BREAKING CHANGE : remove sessionStorage entirely and directly write to disk
