+++ src/utils/JsonFileOperations.ts
@@ -8,9 +8,8 @@ import {
 
 import { Board } from "../interfaces/BoardConfigs";
 import TaskBoard from "main";
-import { loadTasksJsonFromSS } from "./tasksCache";
 
-// Operations with data.json
+// --------------- Operations with data.json ---------------
 
 // Load only the globalSettings part from the data.json
 export const loadGlobalSettings = async (plugin: TaskBoard) => {
@@ -59,13 +58,70 @@ export const saveBoardsData = async (
 	}
 };
 
-// Operations with tasks.json
+// ------------  Operations with tasks.json ----------------
 
+// load tasks from disk.
+export const loadTasksJsonFromDisk = async (
+	plugin: TaskBoard
+): Promise<tasksJson> => {
+	try {
+		const path = `${plugin.app.vault.configDir}/plugins/task-board/tasks.json`;
+		const data: string = await plugin.app.vault.adapter.read(path);
+		const allTasks: tasksJson = JSON.parse(data);
+		return allTasks;
+	} catch (error) {
+		console.error("Error reading tasks.json from disk:", error);
+		throw error;
+	}
+};
+
+// Helper function to clean up the empty entries in tasks.json
+export const dataCleanup = (oldTaskData: tasksJson): tasksJson => {
+	// Function to remove keys with empty arrays from a specified section
+	const removeEmptyKeys = (section: any) => {
+		Object.keys(section).forEach((key) => {
+			if (Array.isArray(section[key]) && section[key].length === 0) {
+				delete section[key];
+			}
+		});
+	};
+
+	// Remove empty arrays from "Pending" and "Completed" sections
+	removeEmptyKeys(oldTaskData.Pending);
+	removeEmptyKeys(oldTaskData.Completed);
+
+	return oldTaskData;
+};
+
+// Function to write tasks data to disk
+export const writeTasksJsonToDisk = async (
+	plugin: TaskBoard,
+	tasksData: tasksJson
+): Promise<void> => {
+	try {
+		const path = `${plugin.app.vault.configDir}/plugins/task-board/tasks.json`;
+
+		if (tasksData) {
+			await plugin.app.vault.adapter.write(
+				path,
+				JSON.stringify(tasksData, null, 4)
+			);
+		} else {
+			console.warn(
+				"Improper tasksData to write to disk."
+			);
+		}
+	} catch (error) {
+		console.warn("Error writing tasks.json to disk:", error);
+	}
+};
+
+// Helper function to load tasks from tasks.json and merge them
 export const loadTasksAndMerge = async (
 	plugin: TaskBoard
 ): Promise<taskJsonMerged> => {
 	try {
-		const allTasks: tasksJson = await loadTasksJsonFromSS(plugin);
+		const allTasks: tasksJson = await loadTasksJsonFromDisk(plugin);
 		const pendingTasks: taskItem[] = [];
 		const completedTasks: taskItem[] = [];
